Use while true instead of waiting for a keyboard input to quit
The function is even worse than just waiting for a keyboard press. The
hard-coded assumption in this lwip_unix_keypressed is that select for
fd 1 will only return on a keyboard but sometimes select just returns
without a timeout or for other reasons and then lwipovpn will just
quit at a random point.